♻️ Refactor: 프로필 수정 뮤테이션 훅 분리 및 아바타 실시간 업데이트 적용
