+++ src/features/user/ui/MyInfoClient.tsx
@@ -1,14 +1,10 @@
 'use client';
 
-import { useQueryClient } from '@tanstack/react-query';
-import { updateMyProfile } from '@/features/user/apis';
+import { useUpdateMyProfileMutation } from '@/features/user/mutations/useUpdateMyProfileMutation';
 import { useMyInfoQuery } from '@/features/user/queries/useMyInfoQuery';
-import type { UpdateMyProfileRequestBody } from '@/features/user/types';
 import ProfileForm from '@/features/user/ui/ProfileForm';
-import { QUERY_KEYS } from '@/shared/constants/queryKey';
 import EmptyState from '@/shared/ui/empty-state/EmptyState';
 import Loading from '@/shared/ui/loading/Loading';
-import { useToastStore } from '@/shared/ui/toast/stores/useToastStore';
 
 /**
  * 내 정보 페이지 클라이언트 컴포넌트
@@ -17,19 +13,8 @@ import { useToastStore } from '@/shared/ui/toast/stores/useToastStore';
  * 서버 컴포넌트인 MyInfoPage에서 이벤트 핸들러를 분리합니다.
  */
 export default function MyInfoClient() {
-  const { showToast } = useToastStore();
-  const queryClient = useQueryClient();
   const { data: user, isLoading, isError, refetch } = useMyInfoQuery();
-
-  const handleUpdateUser = async (body: UpdateMyProfileRequestBody) => {
-    try {
-      await updateMyProfile(body);
-      await queryClient.invalidateQueries({ queryKey: QUERY_KEYS.USER_ME() });
-      showToast('check', '프로필이 저장되었습니다.');
-    } catch {
-      showToast('cancel', '프로필 저장에 실패했습니다. 다시 시도해주세요.');
-    }
-  };
+  const { mutateAsync: updateUser } = useUpdateMyProfileMutation();
 
   return (
     <div className="mt-6 flex w-full flex-col items-center gap-6 md:w-119 lg:w-160">
@@ -44,7 +29,7 @@ export default function MyInfoClient() {
           onRetry={refetch}
         />
       ) : user ? (
-        <ProfileForm user={user} onUpdateUser={handleUpdateUser} />
+        <ProfileForm user={user} onUpdateUser={updateUser} />
       ) : null}
     </div>
   );
